+++ src/components/layout/Layout.tsx
@@ -1,4 +1,3 @@
-import Head from 'next/head';
 import Header from '@/components/layout/headers/Header';
 import Footer from '@/components/layout/footers/Footer';
 import AppScripts from '@/components/AppScripts/AppScripts';
@@ -9,24 +8,6 @@ interface LayoutProps {
 }
 
 const Layout = ({ children }: LayoutProps) => {
-  const PageMetatags = () => {
-    return (
-      <Head>
-        <title>공적인사적모임</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-        <link rel="manifest" href="/site.webmanifest" />
-        <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5" />
-        <meta name="msapplication-TileColor" content="#ffffff" />
-        <meta name="theme-color" content="#ffffff"></meta>
-      </Head>
-    );
-  };
-
   const BackgroundElements = () => {
     return (
       <BackgroundCircleElement>
@@ -40,7 +21,6 @@ const Layout = ({ children }: LayoutProps) => {
 
   return (
     <Wrapper>
-      <PageMetatags />
       <AppScripts />
       <Header />
       <BackgroundElements />
